Brief Overview
Provides a very brief overview of what the repo is and what it contains, since the FiveM docs don't really tell you what this Repo is, just to clone it and move on.

Also added a warning regarding modifying the default resources, and not being allowed to re-release them on the Forums.