Moved JS into HTML file, removed unneeded CSS, removed wanted level
Moved the JS into the HTML file , since a separate file is really not needed.

Move the style CSS into the HTML file, and removed everything else (font included), as it really is overkill, and not required.

Removed wanted level, both from the LUA and HTML/JS, as not once in all my FiveM years have I see this used.

Also swapped out a 31 loop for a `NetworkIsPlayerActive()`.